fix(routers): fixed some backend logic

+++ src/server/api/routers/printers.ts
@@ -1,14 +1,15 @@
 import { z } from "zod";
 import {
     createTRPCRouter,
-    publicProcedure,
+    protectedProcedure,
 } from "~/server/api/trpc";
 
 export const printerRouter = createTRPCRouter({
-    addPrinter: publicProcedure
-        .input(z.object({ userId: z.string(), printerOwner: z.string(), printerBrand: z.string(), printerModel: z.string(), printerType: z.string(), printerArea: z.string() }))
+    addPrinter: protectedProcedure
+        .input(z.object({ printerOwner: z.string(), printerBrand: z.string(), printerModel: z.string(), printerType: z.string(), printerArea: z.string() }))
         .mutation(async ({ input, ctx }) => {
-            const { userId, printerOwner, printerBrand, printerModel, printerType, printerArea } = input;
+            const { printerOwner, printerBrand, printerModel, printerType, printerArea } = input;
+            const userId = ctx.session.user.id;
 
             const newPrinter = await ctx.prisma.printer.create({
                 data: {
@@ -31,10 +32,11 @@ export const printerRouter = createTRPCRouter({
 
             return newPrinter;
         }),
-    getPrinterForSTL: publicProcedure
-        .input(z.object({ bedSize: z.string(), printerType: z.string(), userId: z.string() }))
+    getPrinterForSTL: protectedProcedure
+        .input(z.object({ bedSize: z.string(), printerType: z.string() }))
         .mutation(async ({ input, ctx }) => {
-            const { bedSize, printerType, userId } = input;
+            const { bedSize, printerType } = input;
+            const userId = ctx.session.user.id;
 
             const printers = await ctx.prisma.printer.findMany({
                 where: {
@@ -68,15 +70,13 @@ export const printerRouter = createTRPCRouter({
 
             return printersForSTL;
         }),
-    getMyPrinters: publicProcedure
-        .input(z.object({ userEmail: z.string() }))
-        .query(async ({ input, ctx }) => {
-            const { userEmail } = input;
-
+    getMyPrinters: protectedProcedure
+        .query(async ({ ctx }) => {
+            const userId = ctx.session.user.id;
             const printers = await ctx.prisma.printer.findMany({
                 where: {
                     user: {
-                        email: userEmail,
+                        id: userId,
                     },
                 }
             })
@@ -87,19 +87,34 @@ export const printerRouter = createTRPCRouter({
 
             return printers;
         }),
-    deletePrinter: publicProcedure
+    deletePrinter: protectedProcedure
         .input(z.object({ printerId: z.string() }))
         .mutation(async ({ input, ctx }) => {
             const { printerId } = input;
+            const ownerId = ctx.session.user.id;
+
+            const printerToDelete = await ctx.prisma.printer.findUnique({
+                where: {
+                    id: printerId,
+                },
+            });
+
+            if (!printerToDelete) {
+                throw new Error("No existe la impresora para eliminar");
+            }
+
+            if (printerToDelete.userId !== ownerId) {
+                throw new Error("No tienes permiso para eliminar esta impresora");
+            }
 
             const deletedPrinter = await ctx.prisma.printer.delete({
                 where: {
                     id: printerId,
-                }
-            })
+                },
+            });
 
             if (!deletedPrinter) {
-                throw new Error("No existe la impresora para eliminar");
+                throw new Error("Hubo un problema eliminando la impresora");
             }
 
             return deletedPrinter;
